refactor: move getGithubToken to shared-internal
