+++ jit-binding-server/build.gradle.kts
@@ -14,6 +14,7 @@ dependencies {
     implementation("ch.qos.logback:logback-classic:1.5.6")
 
     implementation(projects.mavenBindingBuilder)
+    implementation(projects.sharedInternal)
 }
 
 application {
